Refactored Dropdown code to separate out Dropdown menu from paddler selection. Fixed bug that incorrectly displayed paddlers in/not in the boat in dropdown.
